+++ raccoon-entities/src/test/java/com/raccoon/entity/repository/ArtistRepositoryTest.java
@@ -2,36 +2,38 @@ package com.raccoon.entity.repository;
 
 import com.raccoon.entity.Artist;
 
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 import javax.inject.Inject;
-import javax.transaction.Transactional;
 
+import io.quarkus.panache.common.Page;
+import io.quarkus.test.TestTransaction;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.h2.H2DatabaseTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 @QuarkusTest
 @QuarkusTestResource(H2DatabaseTestResource.class)
+@TestTransaction
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class ArtistRepositoryTest {
 
     @Inject
     ArtistRepository repository;
 
     @Test
-    @Transactional
-    void findByNameEmpty() {
+    void findByName_should_returnEmpty_when_nameDoesNotExist() {
         var name = "does not exist";
 
-        assertTrue(repository.findByNameOptional(name).isEmpty());
+        assertThat(repository.findByNameOptional(name)).isEmpty();
     }
 
     @Test
-    @Transactional
-    void findByNameOptional() {
+    void findByName_should_returnArtist_when_nameExists() {
         var name = "name";
         var artist = new Artist();
         artist.setName(name);
@@ -39,8 +41,28 @@ class ArtistRepositoryTest {
 
         final var found = repository.findByNameOptional(name);
 
-        assertTrue(found.isPresent());
-        assertEquals(name, found.get().getName());
+        assertThat(found).isPresent();
+        assertThat(found.get().getName()).isEqualTo(name);
+    }
+
+    @Test
+    void listArtistsPaginated_should_returnEmpty_when_noArtistsPresent() {
+        var artists = repository.listArtistsPaginated(Page.of(0, 1));
+
+        assertThat(artists).isEmpty();
+    }
+
+    @Test
+    void listArtistsPaginated_should_returnArtists() {
+        var name = "name";
+        var artist = new Artist();
+        artist.setName(name);
+        repository.persist(artist);
+
+        var artists = repository.listArtistsPaginated(Page.of(0, 1));
+
+        assertThat(artists).hasSize(1);
+        assertThat(artists).contains(artist);
     }
 
 }