+++ raccoon-tests/src/test/java/com/raccoon/RaccoonUserProfileResourceIT.java
@@ -1,6 +1,7 @@
 package com.raccoon;
 
 import com.raccoon.common.ElasticSearchTestResource;
+import com.raccoon.dto.ArtistDto;
 import com.raccoon.entity.Artist;
 import com.raccoon.entity.repository.ArtistReleaseRepository;
 import com.raccoon.entity.repository.ArtistRepository;
@@ -8,7 +9,6 @@ import com.raccoon.entity.repository.UserArtistRepository;
 import com.raccoon.entity.repository.UserRepository;
 import com.raccoon.search.dto.SearchResultArtistDto;
 import com.raccoon.user.UserProfileResource;
-import com.raccoon.user.dto.FollowedArtistDto;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -201,7 +201,7 @@ class RaccoonUserProfileResourceIT {
                 .then()
                 .statusCode(SC_OK);
         // follow the artist
-        FollowedArtistDto artistDto = FollowedArtistDto.builder()
+        ArtistDto artistDto = ArtistDto.builder()
                 .name("name")
                 .spotifyUri("spotifyUri")
                 .lastfmUri("lastfmUri")
@@ -216,12 +216,12 @@ class RaccoonUserProfileResourceIT {
                 .statusCode(SC_NO_CONTENT);
 
         // get followed artists
-        List<FollowedArtistDto> list = given()
+        List<ArtistDto> list = given()
                 .contentType(ContentType.JSON)
                 .when().get("followed-artists")
                 .then()
                 .statusCode(SC_OK)
-                .extract().body().jsonPath().getList("rows", FollowedArtistDto.class);
+                .extract().body().jsonPath().getList("rows", ArtistDto.class);
 
         assertEquals(1, list.size());
         assertEquals(artistDto.getName(), list.get(0).getName());